requirements.txt
streamlit
pandas
python-pptx
Pillow
openpyxl